Clean up for GCC and Intel ICC build.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@5373 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/MdeModulePkg/Bus/Usb/UsbBusDxe/UsbEnumer.c
@@ -720,7 +720,7 @@ UsbEnumerateNewDev (
     //
     if (Parent->Speed == EFI_USB_SPEED_HIGH) {
       Child->Translator.TranslatorHubAddress  = Parent->Address;
-      Child->Translator.TranslatorPortNumber  = Port + 1;
+      Child->Translator.TranslatorPortNumber  = (UINT8) (Port + 1);
 
     } else {
       Child->Translator = Parent->Translator;
